+++ packages/flame/lib/src/components/particle_component.dart
@@ -25,7 +25,6 @@ class ParticleComponent extends Component {
   /// [Particle] within this [Component].
   @override
   void render(Canvas canvas) {
-    super.render(canvas);
     particle.render(canvas);
   }