+++ packages/flame/lib/src/components/sprite_animation_component.dart
@@ -91,7 +91,6 @@ class SpriteAnimationComponent extends PositionComponent with HasPaint {
   @mustCallSuper
   @override
   void render(Canvas canvas) {
-    super.render(canvas);
     animation?.getSprite().render(
           canvas,
           size: size,