+++ packages/flame/lib/src/components/sprite_group_component.dart
@@ -48,7 +48,6 @@ class SpriteGroupComponent<T> extends PositionComponent with HasPaint {
   @mustCallSuper
   @override
   void render(Canvas canvas) {
-    super.render(canvas);
     sprite?.render(
       canvas,
       size: size,